SCRUM-96 fix: 질문글 존재하지 않을 시 예외처리 추가 test: test코드 오류 수정

+++ backend/src/main/java/com/NBE3_4_2_Team4/domain/board/question/controller/QuestionController.java
@@ -6,6 +6,7 @@ import com.NBE3_4_2_Team4.domain.board.question.dto.response.QuestionWriteResDto
 import com.NBE3_4_2_Team4.domain.board.question.entity.Question;
 import com.NBE3_4_2_Team4.domain.board.question.service.QuestionService;
 import com.NBE3_4_2_Team4.domain.member.member.entity.Member;
+import com.NBE3_4_2_Team4.global.exceptions.QuestionNotFoundException;
 import com.NBE3_4_2_Team4.global.rsData.RsData;
 import com.NBE3_4_2_Team4.global.security.AuthManager;
 import com.NBE3_4_2_Team4.standard.dto.PageDto;
@@ -16,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("/api/questions")
@@ -35,11 +37,16 @@ public class QuestionController {
 
     @GetMapping("/{id}")
     public QuestionDto getQuestion(@PathVariable long id) {
-        return new QuestionDto(questionService.findById(id));
+        Question question = questionService.findById(id).orElseThrow(QuestionNotFoundException::new);
+
+        return new QuestionDto(question);
     }
 
     @DeleteMapping("/{id}")
     public RsData<Void> delete(@PathVariable long id) {
+        if (!questionService.findById(id).isPresent()) {
+            throw new QuestionNotFoundException();
+        }
         questionService.delete(id);
         return new RsData<>(
                 "200-1",
@@ -64,13 +71,15 @@ public class QuestionController {
 
     @PutMapping("/{id}")
     @Transactional
-    public RsData<Void> update(@PathVariable long id, @RequestBody @Valid QuestionWriteReqDto reqBody) {
-        Question question = questionService.findById(id);
+    public RsData<QuestionDto> update(@PathVariable long id, @RequestBody @Valid QuestionWriteReqDto reqBody) {
+        Question question = questionService.findById(id).orElseThrow(QuestionNotFoundException::new);
+
         questionService.update(question, reqBody.title(), reqBody.content());
 
         return new RsData<>(
                 "200-1",
-                "%d번 게시글 수정이 완료되었습니다.".formatted(id)
+                "%d번 게시글 수정이 완료되었습니다.".formatted(id),
+                new QuestionDto(question)
         );
     }
 }
